requirements.txt
click
plotly
python-dotenv
